feat(v2): 10x speed gain and candidate class
Read stats from a single game round 10x faster than before. This is
because we're using the stats from the lineup call instead of making a
individual call for it afterwards.

Also add a candidate class that is going to be the basis of the linear
programming problem. This is a work in progress.